requirements.txt
yfinance
pandas
numpy
matplotlib
alpaca-py
python-dotenv
